Fix AG10 doctor health and send refactor residuals
